+++ lib/kafka.rb
@@ -211,6 +211,9 @@ module Kafka
   class FetchError < Error
   end
 
+  class NoPartitionsAssignedError < Error
+  end
+
   # Initializes a new Kafka client.
   #
   # @see Client#initialize