+++ lib/kafka/cluster.rb
@@ -157,6 +157,9 @@ module Kafka
       end
 
       offsets
+    rescue Kafka::ProtocolError
+      mark_as_stale!
+      raise
     end
 
     def resolve_offset(topic, partition, offset)
@@ -190,7 +193,7 @@ module Kafka
     # @raise [ConnectionError] if none of the nodes in `seed_brokers` are available.
     # @return [Protocol::MetadataResponse] the cluster metadata.
     def fetch_cluster_info
-      @seed_brokers.each do |node|
+      @seed_brokers.shuffle.each do |node|
         @logger.info "Fetching cluster metadata from #{node}"
 
         begin